+++ lib/core/widgets/drop_down_floating_label_form_field.dart
@@ -1,14 +1,16 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:test1/core/helper/extensions.dart';
 import 'package:test1/core/theming/my_colors.dart';
 import 'package:test1/core/theming/styles.dart';
 
+import '../../components/search_alert_dialog.dart';
 import '../constants/constants.dart';
 
 class MyDropdownFormField<T> extends StatelessWidget {
   final String labelText;
   final List<T> items;
   final T? value;
+  final String? hint;
   final void Function(T?) onChanged;
   final Widget Function(T) itemBuilder;
   final double? padding;
@@ -24,7 +26,7 @@ class MyDropdownFormField<T> extends StatelessWidget {
     required this.itemBuilder,
     this.padding,
     required this.validator,
-    this.searchHintText = 'ابحث...',
+    this.searchHintText = 'ابحث...', this.hint,
   });
 
   @override
@@ -34,9 +36,10 @@ class MyDropdownFormField<T> extends StatelessWidget {
       child: Padding(
         padding: EdgeInsets.only(bottom: padding ?? 15.0),
         child: GestureDetector(
-          onTap: () => _showSearchableDialog(context),
+          onTap: () => showSearchableDialog(context),
           child: AbsorbPointer(
             child: DropdownButtonFormField<T>(
+              hint: hint!=null?Text(hint!):null,
               style: MyTextStyles.font15Weight600
                   .copyWith(color: MyColors.kPrimaryColor),
               value: value,
@@ -83,8 +86,7 @@ class MyDropdownFormField<T> extends StatelessWidget {
       ),
     );
   }
-
-  void _showSearchableDialog(BuildContext context) {
+  void showSearchableDialog(BuildContext context) {
     showDialog(
       context: context,
       builder: (context) {
@@ -97,103 +99,6 @@ class MyDropdownFormField<T> extends StatelessWidget {
       },
     );
   }
+
 }
 
-class SearchableDropdownDialog<T> extends StatefulWidget {
-  final List<T> items;
-  final Widget Function(T) itemBuilder;
-  final void Function(T?) onChanged;
-  final String searchHintText;
-
-  const SearchableDropdownDialog({
-    super.key,
-    required this.items,
-    required this.itemBuilder,
-    required this.onChanged,
-    this.searchHintText = 'Search...',
-  });
-
-  @override
-  _SearchableDropdownDialogState<T> createState() =>
-      _SearchableDropdownDialogState<T>();
-}
-
-class _SearchableDropdownDialogState<T>
-    extends State<SearchableDropdownDialog<T>> {
-  late List<T> filteredItems;
-  TextEditingController searchController = TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-    filteredItems = widget.items;
-    searchController.addListener(_filterItems);
-  }
-
-  @override
-  void dispose() {
-    searchController.removeListener(_filterItems);
-    searchController.dispose();
-    super.dispose();
-  }
-
-  void _filterItems() {
-    setState(() {
-      filteredItems = widget.items
-          .where((item) => widget
-              .itemBuilder(item)
-              .toString()
-              .toLowerCase()
-              .contains(searchController.text.toLowerCase()))
-          .toList();
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Directionality(
-      textDirection: AppConstants.appTextDirection,
-      child: AlertDialog(
-        title: TextField(
-          controller: searchController,
-          decoration: InputDecoration(
-            enabledBorder: OutlineInputBorder(
-                borderSide: const BorderSide(color: Colors.black, width: 1.5),
-                borderRadius: BorderRadius.circular(20)),
-            focusedBorder: OutlineInputBorder(
-                borderSide: const BorderSide(color: Colors.black, width: 1.5),
-                borderRadius: BorderRadius.circular(20)),
-            focusedErrorBorder: OutlineInputBorder(
-                borderSide: const BorderSide(color: Colors.black, width: 1.5),
-                borderRadius: BorderRadius.circular(20)),
-            hintText: widget.searchHintText,
-            prefixIcon: const Icon(Icons.search),
-          ),
-        ),
-        content: SizedBox(
-          width: double.maxFinite,
-          child: ListView.builder(
-            shrinkWrap: true,
-            itemCount: filteredItems.length,
-            itemBuilder: (context, index) {
-              final item = filteredItems[index];
-              return ListTile(
-                title: widget.itemBuilder(item),
-                onTap: () {
-                  widget.onChanged(item);
-                  Navigator.of(context).pop();
-                },
-              );
-            },
-          ),
-        ),
-        actions: <Widget>[
-          TextButton(
-            onPressed: () => Navigator.of(context).pop(),
-            child: const Text('Close'),
-          ),
-        ],
-      ),
-    );
-  }
-}
